requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-import-export==4.1.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-taggit==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
shortuuid==1.0.13
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
